+++ vcx/libvcx/src/settings.rs
@@ -45,6 +45,7 @@ pub static CONFIG_PROTOCOL_VERSION: &'static str = "protocol_version";
 pub static CONFIG_PAYMENT_METHOD: &'static str = "payment_method";
 pub static CONFIG_TXN_AUTHOR_AGREEMENT: &'static str = "author_agreement";
 pub static CONFIG_POOL_CONFIG: &'static str = "pool_config";
+pub static CONFIG_DID_METHOD: &str = "did_method";
 
 pub static DEFAULT_PROTOCOL_VERSION: usize = 2;
 pub static MAX_SUPPORTED_PROTOCOL_VERSION: usize = 2;