+++ wrappers/rust/indy-sys/src/anoncreds.rs
@@ -218,5 +218,9 @@ extern {
     #[no_mangle]
     pub fn indy_generate_nonce(command_handle: CommandHandle,
                                cb: Option<ResponseStringCB>) -> Error;
+    #[no_mangle]
+    pub fn indy_to_unqualified(command_handle: CommandHandle,
+                               entity: CString,
+                               cb: Option<ResponseStringCB>) -> Error;
 }